get rid of a warning

+++ Source/items.cpp
@@ -1437,7 +1437,7 @@ static void GetStaffSpell(int ii, int lvl)
 	is = &items[ii];
 	sd = &spelldata[bs];
 	static_assert(sizeof(istr) == sizeof(is->_iName), "Mismatching _iName and local string sizes in GetStaffSpell.");
-	if (snprintf(istr, sizeof(istr), "%s of %s", is->_iName, sd->sNameText) >= sizeof(istr))
+	if ((unsigned)snprintf(istr, sizeof(istr), "%s of %s", is->_iName, sd->sNameText) >= sizeof(istr))
 	//if (GetStringWidth(istr) < 125)
 		snprintf(istr, sizeof(istr), "Staff of %s", sd->sNameText);
 	copy_str(is->_iName, istr);
